refactor Sidecar::Interface::InstanceId to separate file
Also, add basic test and doc comments
